requirements.txt
langchain==0.1.20
unstructured 
langchain-core==0.1.52
langchain-google-vertexai==1.0.3
streamlit==1.34.0
streamlit_chat==0.1.1
streamlit_extras==0.4.2
pymongo==4.7.2
google-cloud-secret-manager